+++ lib/interfaces/SpritesheetToSpritesOptions.d.ts
@@ -1,0 +1,9 @@
+import GeneralOptions from './GeneralOptions';
+export default interface SpritesheetToSpritesOptions extends GeneralOptions {
+    /**
+     * The name to prepend to each sprite.
+     *
+     * By default this will use the name of the spritesheet.
+     */
+    name?: string;
+}